Update to 0.0.5
Removed auto-adjust grid. Turns out it was pretty broke. Has something to do  with the canvas updating after a new grid size is set. Will need more investigation.

Corrected the Adjust X/Y buttons so they no longer only adjust the grid in the positive direction. Constant use of these buttons would make the map/grid move to far. Made it so they now adjust in the Positive and Negative directions.

Added a check to see if the drawn grid is 50px or larger. If it is not then user is notified and nothing happens. If it is then numbers show up in the drawn box indicating the size of the square and allows the grid to be set.